fix some linting issues
