+++ lib/rig_inbound_gateway/api_proxy/handler/http.ex
@@ -23,6 +23,7 @@ defmodule RigInboundGateway.ApiProxy.Handler.Http do
 
   alias RigInboundGateway.ApiProxy.Handler
   alias RigInboundGateway.ApiProxy.Handler.HttpHeader
+  alias RigTracing.Context
   alias RigTracing.TracePlug
   @behaviour Handler
 
@@ -45,7 +46,7 @@ defmodule RigInboundGateway.ApiProxy.Handler.Http do
       req_headers
       |> HttpHeader.put_host_header(url)
       |> HttpHeader.put_forward_header(conn.remote_ip, host_ip)
-      |> HttpHeader.put_tracecontext_header(TracePlug.tracecontext_headers())
+      |> TracePlug.put_tracecontext_header(Context.tracecontext())
       |> drop_connection_related_headers()
 
     result = do_request(method, url, body, req_headers)
