+++ lib/rig_inbound_gateway/request_logger/kafka.ex
@@ -5,7 +5,7 @@ defmodule RigInboundGateway.RequestLogger.Kafka do
   use Rig.KafkaConsumerSetup, [:log_topic, :log_schema, :serializer]
 
   alias RigInboundGateway.RequestLogger
-  alias RigTracing.TracePlug
+  alias RigTracing.Context
   @behaviour RequestLogger
   alias UUID
 
@@ -55,7 +55,7 @@ defmodule RigInboundGateway.RequestLogger.Kafka do
           remote_ip: conn.remote_ip |> format_ip
         }
       }
-      |> TracePlug.append_distributed_tracing_context(TracePlug.tracecontext_headers())
+      |> Context.append_tracecontext(Context.tracecontext())
       |> Poison.encode!()
 
     produce("partition", kafka_message)
